refactoring round 3
